@@ -1,31 +1,29 @@
 import React from "react";
-import { Link, Route, Switch, useRouteMatch } from "react-router-dom";
+import { Link, Route, Routes } from "react-router-dom";
 import UserProfile from "./userprofile";
 import FarmManagement from "./farmmanagement";
 import TaskAssignment from "./taskassignment";
 import TaskProgress from "./taskprogress";
 
 const Dashboard = () => {
-  const { path, url } = useRouteMatch();
-
   return (
     <div>
       <nav>
         <ul>
           <li>
-            <Link to={`${url}`}>Dashboard</Link>
+            <Link to="/">Dashboard</Link>
           </li>
           <li>
-            <Link to={`${url}/user_profile`}>My Profile</Link>
+            <Link to="/user_profile">My Profile</Link>
           </li>
           <li>
-            <Link to={`${url}/farm_management`}>Farm Management</Link>
+            <Link to="/farm_management">Farm Management</Link>
           </li>
           <li>
-            <Link to={`${url}/task_assignment`}>Task Assignment</Link>
+            <Link to="/task_assignment">Task Assignment</Link>
           </li>
           <li>
-            <Link to={`${url}/task_progress`}>Task Progress</Link>
+            <Link to="/task_progress">Task Progress</Link>
           </li>
           <li>
             <Link to="/logout">Logout</Link>
@@ -33,15 +31,13 @@ const Dashboard = () => {
         </ul>
       </nav>
       <div>
-        <Switch>
-          <Route exact path={path}>
-            <h2>Welcome to the Dashboard!</h2>
-          </Route>
-          <Route path={`${path}/user_profile`} component={UserProfile} />
-          <Route path={`${path}/farm_management`} component={FarmManagement} />
-          <Route path={`${path}/task_assignment`} component={TaskAssignment} />
-          <Route path={`${path}/task_progress`} component={TaskProgress} />
-        </Switch>
+        <Routes>
+          <Route path="/" element={<h2>Welcome to the Dashboard!</h2>} />
+          <Route path="/user_profile" element={<UserProfile />} />
+          <Route path="/farm_management" element={<FarmManagement />} />
+          <Route path="/task_assignment" element={<TaskAssignment />} />
+          <Route path="/task_progress" element={<TaskProgress />} />
+        </Routes>
       </div>
     </div>
   );
