@@ -1,12 +1,15 @@
 import { jsxDEV } from "react/jsx-dev-runtime";
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
 import "./App.css";
 
-ReactDOM.render(
+const root = document.getElementById("root");
+
+const reactRoot = createRoot(root);
+
+reactRoot.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
