@@ -12,8 +12,8 @@ import os
 from flask_cors import CORS
 
 app = Flask(__name__)
+CORS(app)
 jwt = JWTManager(app)
-CORS(app, resources={r"/api/*": {"origins": "http://localhost:5173"}})
 
 
 flask_secret_key = secrets.token_hex(16)
